+++ example/expo-router-example/app/_layout.tsx
@@ -9,7 +9,7 @@ import { SplashScreen, Stack } from 'expo-router';
 import React, { useEffect } from 'react';
 import { useColorScheme } from 'react-native';
 import { StyledProvider } from '@gluestack-style/react';
-import { config } from '../gluestack-ui.config';
+import { config } from '@gluestack-ui/config';
 
 export {
   // Catch any errors thrown by the Layout component.