+++ packages/themed/src/components/ImageBackground/styled-components/Root.tsx
@@ -1,0 +1,6 @@
+import { ImageBackground } from 'react-native';
+import { styled } from '@gluestack-style/react';
+
+export default styled(ImageBackground, {}, {
+  componentName: 'ImageBackground',
+} as const);